// src/app/api-client/model/licenseObligation.ts
/**
 * Horde Model Reference API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * Represents a common obligation that consumers may need to surface.
 */
export const LicenseObligation = {
  Attribution: 'attribution',
  IncludeLicense: 'include_license',
  ShareAlike: 'share_alike',
  DiscloseSource: 'disclose_source',
  NetworkSourceDisclosure: 'network_source_disclosure',
  ResponsibleUseRestrictions: 'responsible_use_restrictions',
  Other: 'other',
} as const;
export type LicenseObligation = (typeof LicenseObligation)[keyof typeof LicenseObligation];